Move staticlint.js to a separete file
