+++ lib/index.js
@@ -14,7 +14,6 @@ let ignoreInfo;
 let ignoreWarning;
 let showErrorCodes;
 let ignoreIssueCodes;
-let JLC = null;
 
 export function activate() {
   // eslint-disable-next-line global-require
@@ -65,46 +64,11 @@ export function deactivate() {
   }
 }
 
-
-export function provideLinterStaticLint() {
-  // Using atom-languageclient
-  const cp = require('child_process');
-  const path = require('path');
-  const { AutoLanguageClient } = require('atom-languageclient');
-
-  class JuliaLanguageClient extends AutoLanguageClient {
-    getGrammarScopes() {
-      return ['source.julia'];
-    }
-
-    getLanguageName() {
-      return 'Julia';
-    }
-
-    getServerName() {
-      return 'Juno';
-    }
-
-    startServerProcess() {
-      return cp.spawn(executablePath, [path.resolve(__dirname, '..', 'script', 'StaticLint-juia-server.jl')]);
-    }
-  }
-
-  JLC = new JuliaLanguageClient();
-  JLC.config = {
-    jlpath: {
-      type: 'string',
-      default: 'julia',
-      name: 'Path',
-      description: 'The location of the Julia binary.',
-      order: 1,
-    },
-  };
-  module.exports = JLC;
-
-  return {
-  };
-}
+// moved to staticlint.js
+// export function provideLinterStaticLint() {
+//   return {
+//   };
+// }
 
 export function provideLinterLint() {
   return {
@@ -162,9 +126,10 @@ export function provideLinterLint() {
 
 
 export function provideLinter() {
-  if (backend === 'Lint') {
-    return provideLinterLint();
-  } else if (backend === 'StaticLint') {
-    return provideLinterStaticLint();
-  }
+  return provideLinterLint();
+  // if (backend === 'Lint') {
+  //   return provideLinterLint();
+  // } else if (backend === 'StaticLint') {
+  //   return provideLinterStaticLint();
+  // }
 }
